requirements.txt
google-generativeai
reportlab
jsonschema
pytesseract
Pillow
PyPDF2